Sesame Street data based on Hox (2010)

Synthetic data based on Hox (2010, p. 16). In the study, the outcome variable popular represents the popularity score of pupils, ranging from 0 (very unpopular) to 10 (very popular), for pupils nested in 100 classes of varying size. The popularity scores are predicted by pupil level predictors gender (G) and pupil extroversion scores (PE) that range from 1 (introversion) to 10 (extroversion), a class-level predictor teacher experience (TE), and the cross-level interactions between G and TE as well as PE and TE. Since standardization is recommended when the model contains interactions, we standardize PS, PE and TE by means of grand mean centering. That is, we first subtract the overall means of the continuous variables PS, PE, and TE from each of their values, before dividing these values by their standard deviations.

data(hox_2010)

Format

A data frame with 2000 rows and 6 variables.

Details

ID	integer	Pupil ID

class	integer	Class ID

PE	numeric	Pupil extroversion, standardized

G	factor	Pupil sex

PS	numeric	Popularity scores, standardized

TE	integer	Teacher experience, standardized

References

Hox, J. J. (2010). Multilevel analysis: Techniques and applications (2nd ed.). New York, NY: Routledge.