Wechsler intelligence test data

Dataset based on McArdle and Prescott (1992, p.90). This study evaluates intelligence and cognitive ability in a sample of individuals over 18 years of age (N = 1680) using the IQ test Wechsler Adult Intelligence Scale-Revised (WAIS-R) (Wechsler, 1981).

data(wechsler)

Format

A data frame with 1680 rows and 10 variables.

Details

age	integer	Participants' age (recoded)

edc	factor	Whether a participant graduated high school or not (1 = not graduated, 2 = graduated)

y1	integer	information; general knowledge of participants

y2	integer	comprehension; ability of abstract reasoning or judgment

y3	integer	similarities; unifying a theme

y4	integer	vocabulary; verbal definition

y5	integer	picture completion; perceiving visual images with missing features

y6	integer	block design; arranging blocks to match a design

y7	integer	picture arrangement; ordering cards with true story lines

y8	integer	object assembly; reassembling puzzles

References

McArdle, J. J., & Prescott, C. A. (1992). Age-based construct validation using structural equation modeling. Experimental Aging Research, 18, 87-115.